Smaller fixes
