+++ client/components/Filter.tsx
@@ -56,11 +56,6 @@ export const InfoSourceFilter: React.FC<{
     );
 }
 
-/**
- * TODO:
- * - Extract tags to component
- * - Move it to a Modal
-*/
 export const FilterMenu: React.FC<{ onClose: () => void }> = ({ onClose }) => {
     const { tags: allTags } = useTagContext();
     const { filter, setFilter } = useGamesContext();
